Data process/requirements.txt
moviepy
python-dotenv
boto3
supabase 
loguru
requests
pytest
ipykernel